[RUNTIME][VM] Lift LoadVMModule() interface to VM runtime
